# backend/internal/testapi/server.gen.go
// Package testapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package testapi

import (
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for SeedUserDataRole.
const (
	SeedUserDataRoleAdmin        SeedUserDataRole = "admin"
	SeedUserDataRoleBrandManager SeedUserDataRole = "brand_manager"
)

// Valid indicates whether the value is a known member of the SeedUserDataRole enum.
func (e SeedUserDataRole) Valid() bool {
	switch e {
	case SeedUserDataRoleAdmin:
		return true
	case SeedUserDataRoleBrandManager:
		return true
	default:
		return false
	}
}

// Defines values for SeedUserRequestRole.
const (
	SeedUserRequestRoleAdmin        SeedUserRequestRole = "admin"
	SeedUserRequestRoleBrandManager SeedUserRequestRole = "brand_manager"
)

// Valid indicates whether the value is a known member of the SeedUserRequestRole enum.
func (e SeedUserRequestRole) Valid() bool {
	switch e {
	case SeedUserRequestRoleAdmin:
		return true
	case SeedUserRequestRoleBrandManager:
		return true
	default:
		return false
	}
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error struct {
		Code    string `json:"code"`
		Message string `json:"message"`
	} `json:"error"`
}

// ResetTokenData defines model for ResetTokenData.
type ResetTokenData struct {
	Token string `json:"token"`
}

// ResetTokenResult defines model for ResetTokenResult.
type ResetTokenResult struct {
	Data ResetTokenData `json:"data"`
}

// SeedBrandData defines model for SeedBrandData.
type SeedBrandData struct {
	Id   string `json:"id"`
	Name string `json:"name"`
}

// SeedBrandRequest defines model for SeedBrandRequest.
type SeedBrandRequest struct {
	// ManagerEmail If provided, assigns this user as manager (creates user if needed)
	ManagerEmail *openapi_types.Email `json:"managerEmail,omitempty"`
	Name         string               `json:"name"`
}

// SeedBrandResult defines model for SeedBrandResult.
type SeedBrandResult struct {
	Data SeedBrandData `json:"data"`
}

// SeedUserData defines model for SeedUserData.
type SeedUserData struct {
	Email openapi_types.Email `json:"email"`
	Id    string              `json:"id"`
	Role  SeedUserDataRole    `json:"role"`
}

// SeedUserDataRole defines model for SeedUserData.Role.
type SeedUserDataRole string

// SeedUserRequest defines model for SeedUserRequest.
type SeedUserRequest struct {
	Email    openapi_types.Email `json:"email"`
	Password string              `json:"password"`
	Role     SeedUserRequestRole `json:"role"`
}

// SeedUserRequestRole defines model for SeedUserRequest.Role.
type SeedUserRequestRole string

// SeedUserResult defines model for SeedUserResult.
type SeedUserResult struct {
	Data SeedUserData `json:"data"`
}

// GetResetTokenParams defines parameters for GetResetToken.
type GetResetTokenParams struct {
	Email openapi_types.Email `form:"email" json:"email"`
}

// SeedBrandJSONRequestBody defines body for SeedBrand for application/json ContentType.
type SeedBrandJSONRequestBody = SeedBrandRequest

// SeedUserJSONRequestBody defines body for SeedUser for application/json ContentType.
type SeedUserJSONRequestBody = SeedUserRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Get the raw reset token captured for the given email
	// (GET /test/reset-tokens)
	GetResetToken(w http.ResponseWriter, r *http.Request, params GetResetTokenParams)
	// Create a test brand and optionally assign a manager
	// (POST /test/seed-brand)
	SeedBrand(w http.ResponseWriter, r *http.Request)
	// Create a test user with given credentials and role
	// (POST /test/seed-user)
	SeedUser(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Get the raw reset token captured for the given email
// (GET /test/reset-tokens)
func (_ Unimplemented) GetResetToken(w http.ResponseWriter, r *http.Request, params GetResetTokenParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Create a test brand and optionally assign a manager
// (POST /test/seed-brand)
func (_ Unimplemented) SeedBrand(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Create a test user with given credentials and role
// (POST /test/seed-user)
func (_ Unimplemented) SeedUser(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetResetToken operation middleware
func (siw *ServerInterfaceWrapper) GetResetToken(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetResetTokenParams

	// ------------- Required query parameter "email" -------------

	if paramValue := r.URL.Query().Get("email"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "email"})
		return
	}

	err = runtime.BindQueryParameterWithOptions("form", true, true, "email", r.URL.Query(), &params.Email, runtime.BindQueryParameterOptions{Type: "string", Format: "email"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "email", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetResetToken(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// SeedBrand operation middleware
func (siw *ServerInterfaceWrapper) SeedBrand(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.SeedBrand(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// SeedUser operation middleware
func (siw *ServerInterfaceWrapper) SeedUser(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.SeedUser(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/test/reset-tokens", wrapper.GetResetToken)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/test/seed-brand", wrapper.SeedBrand)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/test/seed-user", wrapper.SeedUser)
	})

	return r
}